requirements.txt
pdf2image
streamlit
pdfminer.six
Pillow
pandas
